refact: rename "display" to "screen"
