+++ src/view/screen/PokemonAbilities.ts
